Move season embed to helper
